Hook updates
Removed the hook trigger for OnEntityAttacked from BasePlayer.Hurt as it
is also called from BaseCombatEntity.Hurt for a BasePlayer. When a
player was taking damage it was triggering the hook twice.
Renamed the hook OnEntityAttacked to OnEntityTakeDamage as it is a
pre-hook. A wrapper has been added to prevent plugins using the old hook
from breaking.
Updated the IL index for OnEntityTakeDamage.
Added the hook OnPlayerSleepEnded(BasePlayer player)
Added the hook OnLoseCondition(Item item, ref float amount). Currently
the hook is of not much use in plugins written in other languages than
C# as it is a pre-hook but the value can't be editted from other plugin
languages.
Update RustCore.cs to add summaries for all the wrappers that have been
added recently.