Fixes, changes, new hook
Corrected the injection index for the hooks OnEntitySpawned and
OnSignUpdated.
Renamed the hook CanOpenDoor to CanUseDoor. A wrapper has been added to
prevent plugins that are using CanOpenDoor from breaking.
Added OnEntityGroundMissing(BaseEntity entity). This hook is triggered
on entities (sleepingbags, furnaces, signs, lanterns, storage
crates,...) when the foundation underneath it (or wall behind it) is
removed. Returning a value will cancel the removal of the entity.