[TheForest] Add standard and Covalence player login hooks
This required a manual patch to make CoopKickToken
public instead of internal. The Oxide Patcher will handle this
eventually, but for now just keep it in mind!

Standard:
- object CanClientLogin(BoltConnection connection)
- void OnUserApprove(BoltConnection connection)

Covalence:
- object CanUserLogin(string name, string id, string ip)
- void OnUserApproved(string name, string id, string ip)

The IP currently resolves to 1.16.0.1 though as The Forest doesn't
seem to have it set to a valid IP address, but looking into it.